+++ tests/testthat/test-biofilter_toc.R
@@ -76,88 +76,37 @@ test_that("biofilter_toc correctly handles temperatures and ozonated water.", {
 ################################################################################*
 ################################################################################*
 # biofilter_toc helpers ----
-test_that("biofilter_toc_once outputs are the same as base function, biofilter_toc", {
-  water1 <- suppressWarnings(define_water(7.9, 20, 50,
-    tot_hard = 50, ca = 13,
-    na = 20, k = 20, cl = 30, so4 = 20,
-    tds = 200, cond = 100,
-    toc = 2, doc = 1.8, uv254 = 0.05, br = 50
-  )) %>%
-    biofilter_toc(ebct = 10)
-
-  water2 <- suppressWarnings(water_df %>%
-    slice(1) %>%
-    define_water_chain() %>%
-    biofilter_toc_once(ebct = 10))
-
-  expect_equal(water1@doc, water2$doc)
-})
-
-# Check that output is a data frame
-
-test_that("biofilter_toc_once is a data frame", {
-  water1 <- suppressWarnings(water_df %>%
-    slice(1) %>%
-    define_water_chain() %>%
-    balance_ions_chain() %>%
-    biofilter_toc_once(
-      input_water = "balanced_water",
-      ebct = 5
-    ))
-
-  expect_true(is.data.frame(water1))
-})
-
-# Check biofilter_toc_once can use a column or function argument for ebct
-
-test_that("biofilter_toc_once can use a column or function argument for ebct", {
-  water1 <- suppressWarnings(water_df %>%
-    slice(1) %>%
-    define_water_chain() %>%
-    balance_ions_chain() %>%
-    biofilter_toc_once(
-      input_water = "balanced_water",
-      ebct = 5
-    ))
-  water2 <- suppressWarnings(water_df %>%
-    slice(1) %>%
-    define_water_chain() %>%
-    mutate(
-      ebct = 5
-    ) %>%
-    balance_ions_chain() %>%
-    biofilter_toc_once(input_water = "balanced_water"))
-
-  water3 <- suppressWarnings(water_df %>%
-    slice(1) %>%
-    define_water_chain() %>%
-    mutate(ebct = 5) %>%
-    balance_ions_chain() %>%
-    biofilter_toc_once(input_water = "balanced_water", ozonated = TRUE))
-
-  expect_equal(water1$doc, water2$doc) # test different ways to input args
-  # Test that inputting ebct and ozonated separately (in column and as an argument) gives same results
-  expect_equal(water1$doc, water3$doc)
-})
 
 test_that("biofilter_toc_chain outputs are the same as base function, biofilter_toc", {
-  water1 <- suppressWarnings(define_water(7.9, 20, 50,
-    tot_hard = 50, ca = 13,
+  water0 <- define_water(7.9, 20, 50,
+    tot_hard = 50, ca = 13, mg = 4,
     na = 20, k = 20, cl = 30, so4 = 20,
     tds = 200, cond = 100,
     toc = 2, doc = 1.8, uv254 = 0.05
-  ) %>%
-    biofilter_toc(ebct = 10))
+  )
 
-  water2 <- suppressWarnings(water_df %>%
+  water1 <- biofilter_toc(water0, ebct = 10)
+
+  water2 <- water_df %>%
     slice(1) %>%
     define_water_chain() %>%
     biofilter_toc_chain(ebct = 10, output_water = "biof") %>%
-    pluck_water("biof", c(
-      "doc"
-    )))
+    pluck_water("biof", c("doc"))
+
+  ebcts <- tibble(EBCT = c(10, 20, 30))
+  ozone <- tibble(ozonated = c(T, F))
+  water3 <- water_df %>%
+    slice(1) %>%
+    define_water_chain("raw") %>%
+    cross_join(ebcts) %>%
+    cross_join(ozone) %>%
+    biofilter_toc_chain("raw", "biof", ebct = EBCT) %>%
+    pluck_water("biof", c("doc"))
+
+  water4 <- biofilter_toc(water0, ebct = 20, ozonated = F)
 
   expect_equal(water1@doc, water2$biof_doc)
+  expect_equal(water4@doc, water3$biof_doc[4])
 })
 
 # Test that output is a column of water class lists, and changing the output column name works